+++ api/src/generator/entities/entity/attribute.template.ts
