+++ api/src/generator/entities/entity/entity.template.ts
