+++ api/src/generator/entities/entity/index.ts
@@ -1,0 +1,1 @@
+export * from './entity';
